requirements-dev.txt
attrs==22.1.0
black==22.10.0
boto3-stubs==1.26.10
boto3-stubs-lite==1.26.10
botocore-stubs==1.29.9
isort==5.10.1
mypy-boto3-cloudformation==1.26.0.post1
mypy-boto3-dynamodb==1.26.0.post1
mypy-boto3-ec2==1.26.7
mypy-boto3-lambda==1.26.0.post1
mypy-boto3-rds==1.26.9
mypy-boto3-rds-data==1.26.0.post1
mypy-boto3-s3==1.26.0.post1
mypy-boto3-sqs==1.26.0.post1
mypy-boto3-sts==1.26.0.post1
mypy-extensions==0.4.3
pycodestyle==2.9.1
pydocstyle==6.1.1
pyflakes==2.5.0
pylama==8.4.1
pyparsing==3.0.9
pytest==7.2.0
yawsso==1.0.1
